refactor: split dashboard into a package (eval / validate / shared loop)
The single dashboard.py grew to hold both the eval and validate views. Split it into a
`cli/dashboard/` package: `eval.py` and `validate.py` each own their view, `base.py` holds the
one substantial shared piece (the `live_view` refresh loop), and `__init__` re-exports
`dashboard` / `validate_dashboard` / `TaskProgress` so callers are unchanged. Each view builds
its own (small, differing) layout grids inline rather than behind shared micro-helpers.
